cloudinary updates

+++ controllers/postController.js
@@ -5,6 +5,11 @@ const Reply = require('../Models/post').Reply;
 const Share = require('../Models/post').Share;
 const User = require('../Models/user');
 const upload = require("../middlewares/uploadFile");
+const path = require("path");
+const fs = require("fs");
+const { cloudinaryUploadImage, cloudinaryRemoveImage, getImageUrl } = require("../Services/cloudinary");
+
+
 
 //create post
 const addPost = async (req, res, next) => {
@@ -16,11 +21,18 @@ const addPost = async (req, res, next) => {
         const newPost = new Post({
             author: author, 
             content: content,
-            dateCreated : Date.now(),
-            attachments: files.map(file => ({ 
-                type: file.mimetype.split('/')[0], 
-                url: file.path // Store the file path as the URL (you may need to adjust this based on your file storage setup)
-            }))
+            attachments: files.map(async file => { 
+                // Upload the file to cloudinary
+                const imageUrl = await getImageUrl(file.path);
+                return {
+
+                    type: file.mimetype.split('/')[0], 
+                    url: imageUrl
+                    
+                }
+                
+                
+            })
         });
 
         await newPost.save();
@@ -33,7 +45,7 @@ const addPost = async (req, res, next) => {
         }
         next(err);
     }
-}
+};
 
 //update post
 const updatePost = async (req, res, next) => {
@@ -41,7 +53,6 @@ const updatePost = async (req, res, next) => {
         //find the id of the post to be updated
         const post = await Post.findById(req.params.id);
         await post.updateOne({$set:req.body});
-
         res.status(200).json({message:'The post has been successfully updated'})
     }
     catch(err){
@@ -51,7 +62,7 @@ const updatePost = async (req, res, next) => {
         }
         next(err);  
     }
-}
+};
 
 //delete post
 const deletePost = async (req, res, next) => {
@@ -68,7 +79,19 @@ const deletePost = async (req, res, next) => {
       }
       next(err);  
     }
-}
+};
+
+//get all posts
+const getAllPosts = async (req, res) => {
+    try {
+      const posts = await Post.find();
+      console.log(posts);
+      res.status(200).json(posts);
+    } catch (error) {
+      console.error(error);
+      res.status(500).json({ error: 'Internal Server Error' });
+    }
+};
 
 //react post
 const reactPost = async (req, res, next) => {
@@ -121,7 +144,6 @@ const addComment  = async (req,res,next)=>{
         const newComment = new  Comment ({
             author: author,
             content:content,
-            dateCreated : Date.now(),
             attachments: files.map(file => ({
                 type: file.mimetype.split('/')[0], 
                 url: file.path
@@ -141,11 +163,40 @@ const addComment  = async (req,res,next)=>{
     }
 };
 
+//update comment
+const updateComment = async (req, res, next) => {
+    try{
+        // Find the post to which the comment belongs
+        const comment = await Comment.findById(req.params.id);
+        await comment.updateOne({$set:req.body});
+        res.status(200).json({message:'The comment has been successfully updated'})
+    }
+    catch(err){
+        console.error('Error updating post.', err);
+        if (!err.statusCode) {
+            err.statusCode = 500;
+        }
+        next(err);  
+    }
+};
+
+//delete comment
 const deleteComment = async (req, res, next) => {
     try{
-        //find the id of the post to be deleted
-        const comment = await Comment.findById(req.params.id);
+        // Find the post to which the comment belongs
+        const post = await Post.findById(req.params.postId);
+        if (!post) {
+            return res.status(404).json({ message: "Post not found" });
+        }
+        //find the id of the comment to be deleted
+        const comment = await Comment.findById(req.params.commentId);
+        if (!comment) {
+            return res.status(404).json({ message: "Comment not found" });
+        }
+        
+        post.comments.pull(comment);
         await comment.deleteOne();
+        await post.save();
         res.status(200).json({message:'Your comment has been successfully deleted.'})
     }
     catch(err){
@@ -155,7 +206,8 @@ const deleteComment = async (req, res, next) => {
       }
       next(err);  
     }
-}
+};
+
 //react for a comment
 const reactComment = async (req, res, next) => {
     try {
@@ -196,20 +248,35 @@ const reactComment = async (req, res, next) => {
 
 //reply for a comment
 const replyComment  = async (req,res,next)=>{
-    try{    
-        const comment = await Comment.findById(req.params.id);
+    try{   
+        // Find the post to which the comment belongs
+        const post = await Post.findById(req.params.postId);
+        if (!post) {
+            return res.status(404).json({ message: "Post not found" });
+        }
+        const comment = await Comment.findById(req.params.commentId);
         if (!comment) {
             return res.status(404).json({ error: 'This comment does not exist' });
         }
+        const  files = req.files;
+        const {author, content} = req.body;
 
-        const  reply = new Reply (req.body);
+        const  reply = new Reply ({
+            author: author,
+            content: content,
+            attachments: files.map(file => ({
+                type: file.mimetype.split('/')[0],
+                url: file.path,
+            })),
+        });
+        comment.replies = comment.replies || []; // Initialize replies if it doesn't exist
 
-        comment.replies = comment.replies || {}; // Initialize replies if it doesn't exist
-        
         comment.replies.push(reply);
+        comment.count += 1;
+        
         await reply.save(); 
         await comment.save();  
-        
+        await post.save();
         res.status(200).json({ message: 'Your reply was shared.' });
 
     } catch (err) {
@@ -221,10 +288,80 @@ const replyComment  = async (req,res,next)=>{
     }
 };
 
+//react for a reply
+const reactReply = async (req, res, next) => {
+    try {
+        const comment = await Comment.findById(req.params.commentId);
+        if (!comment) {
+            return res.status(404).json({ error: 'This comment does not exist' });
+        }
+
+        const reply = await Reply.findById(req.params.replyId);
+        if (!reply) {
+            return res.status(404).json({ error: 'This reply does not exist' });
+        }
+
+      //Check if the user has already reacted to the comment
+      let userReacted = reply.reacts.users.includes(req.body.author);
+      if (! userReacted) {
+        // User has not reacted, allow reacting to the comment
+        reply.reacts.count += 1;
+        reply.reacts.users.push(req.body.author);
+        userReacted = true;
+        res.status(200).json({message:'You liked the comment.'});
+      } else {
+        // User has already reacted, allow disliking the comment
+        reply.reacts.count -= 1;
+        reply.reacts.users = reply.reacts.users.pull(req.body.author);
+        userReacted = false;
+        res.status(200).json({message:'You disliked the comment.'});
+      }
+
+      await reply.save();
+      await comment.save();
+      res.status(200).json({ message: 'Your comment was reacted.' });
+
+    } catch (err) {
+        console.error("Error reacting comment.", err);
+        if (!err.statusCode) {
+            err.statusCode = 500;
+        }
+        next(err);
+    }
+};
+
+//delete reply
+const deleteRely = async (req, res, next) => {
+    try{
+        
+        const comment = await Comment.findById(req.params.commentId);
+        if (!comment) {
+            return res.status(404).json({ message: "Comment not found" });
+        }
+
+        const reply = await Reply.findById(req.params.replyId);
+        if (!reply) {
+            return res.status(404).json({ message: "Reply not found" });
+        }
+        
+        comment.replies.pull(reply);
+        await reply.deleteOne();
+        await comment.save();
+        res.status(200).json({message:'Your comment has been successfully deleted.'})
+    }
+    catch(err){
+      console.error('Error deleting post.', err);
+      if (!err.statusCode) {
+        err.statusCode = 500;
+      }
+      next(err);  
+    }
+};
+
 //save post
 const savePosts = async(req, res, next) =>{
     try{
-        const user = await User.findById(req.body);
+        const user = await User.findById(req.body.author);
         if (!user) {
             return res.status(404).json({ error: 'User not found' });
         }
@@ -235,14 +372,14 @@ const savePosts = async(req, res, next) =>{
         }
 
         //Check if the user has already saved the post
-        const isSaved = User.savedPosts.includes(req.params.id);
-        if (isSaved) {
-            user.savedPosts.push(req.params.id);
+        const isSaved = User.savePosts.includes(req.params.id);
+        if (!isSaved) {
+            user.savePosts.push(req.params.id);
             await user.save();
             res.status(200).json({ message: 'You saved the post.' });
         } else{
             // Remove the post ID from savedPosts to unsave it
-            user.savedPosts = user.savedPosts.filter((postId) => postId !== req.params.id);
+            user.savePosts = user.savePosts.filter((postId) => postId !== req.params.id);
             await user.save();
             res.status(200).json({ message: 'You unsaved the post.' });
         }
@@ -273,10 +410,10 @@ const  sharePost = async (req, res, next) => {
           post.shares = post.shares || {}; 
           post.shares.count = post.shares.count || 0;
 
+          post.shares.users = post.shares.users || []; // Initialize users array if it doesn't exist
           post.shares.count += 1;
           post.shares.users.push(req.body.author);
-        
-          //
+         
           await sharedPost.save();
           await post.save();
 
@@ -296,11 +433,15 @@ module.exports = {
     addPost,
     updatePost,
     deletePost,
-    addComment,
-    deleteComment,
+    getAllPosts,
     reactPost,
+    addComment,
+    updateComment,
+    deleteComment,
     reactComment,
     replyComment,
+    reactReply,
+    deleteRely,
     savePosts,
     sharePost,
   };
